[Move] Add src/components/community folder and move PostForm, PostList component to this folder
